refactor(db): extract /authors and /series index pages to db/src/browse.rs
